requirements.txt
flask
flask_pymongo
python-dotenv
flask_bcrypt
pyjwt
pydantic
email-validator
flask_cors
PyJWT